6-12 토큰 유효성 적용

+++ src/routes/board/board.service.ts
@@ -1,4 +1,9 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 
@@ -16,58 +21,6 @@ export class BoardService {
     private boardRepository: Repository<Board>,
   ) {}
 
-  private boards = [
-    {
-      name: 'Inez Dooley',
-      contents: 'contents 1',
-      id: 1,
-    },
-    {
-      name: 'Mrs. Bob Brown',
-      contents: 'contents 2',
-      id: 2,
-    },
-    {
-      name: 'Sheila White',
-      contents: 'contents 3',
-      id: 3,
-    },
-    {
-      name: 'Mindy Ruecker',
-      contents: 'contents 4',
-      id: 4,
-    },
-    {
-      name: 'Nelson Schowalter',
-      contents: 'contents 5',
-      id: 5,
-    },
-    {
-      name: 'Debra Armstrong PhD',
-      contents: 'contents 6',
-      id: 6,
-    },
-    {
-      name: 'Deanna Bailey',
-      contents: 'contents 7',
-      id: 7,
-    },
-    {
-      name: 'Misty Connelly',
-      contents: 'contents 8',
-      id: 8,
-    },
-    {
-      name: 'Kim Ruecker',
-      contents: 'contents 9',
-      id: 9,
-    },
-    {
-      name: 'Sophia VonRueden',
-      contents: 'contents 10',
-      id: 10,
-    },
-  ];
 
   async findAll() {
     return this.boardRepository.find();
@@ -85,7 +38,7 @@ export class BoardService {
     return this.boardRepository.save(data);
   }
 
-  async update(id: number, data: UpdateBoardDto) {
+  async update(userId : number ,id: number, data: UpdateBoardDto) {
     console.log(id, data);
     
     const board = await this.boardRepository.findOneBy({
@@ -95,16 +48,23 @@ export class BoardService {
     console.log(board);
 
     if (!board) throw new HttpException('NOT_FOUND', HttpStatus.NOT_FOUND);
-
+    
+    if (userId !== board.userId) {
+      throw new UnauthorizedException();
+    }
     return this.boardRepository.update(id, {
       ...data
     })
   }
 
-  async delete(id: number) {
+  async delete(userId: number, id: number) {
     const board = await this.getBoardById(id);
 
     if (!board) throw new HttpException('NOT_FOUND', HttpStatus.NOT_FOUND);
+
+    if (userId !== board.userId) {
+      throw new UnauthorizedException();
+    }
 
     return this.boardRepository.remove(board);
   }
